LayoutTests/imported/w3c/web-platform-tests/fetch/api/cors/cors-cookies.any-expected.txt

CONSOLE MESSAGE: line 46: ReferenceError: Can't find variable: get_host_info

Harness Error (FAIL), message = ReferenceError: Can't find variable: get_host_info
